Move format_duration to helpers as a generic utility
